move prisma folder back to main app directory
