+++ db/access/game.ts
@@ -1,5 +1,5 @@
 import { Game, Prisma } from '@prisma/client';
-import { platformDB } from '../prisma';
+import { platformDB } from '../../prisma';
 
 export async function createGame(gameData: Prisma.GameCreateArgs['data']): Promise<Game> {
   return platformDB.game.create({
