+++ db/access/head-coach-game.ts
@@ -1,5 +1,5 @@
 import { HeadCoachGame, Prisma } from '@prisma/client';
-import { platformDB } from '../prisma';
+import { platformDB } from '../../prisma';
 
 export async function createHeadCoachGame(
   headCoachGameData: Prisma.HeadCoachGameCreateArgs['data']
