+++ db/access/official-game.ts
@@ -1,5 +1,5 @@
 import { Prisma } from '@prisma/client';
-import { platformDB } from '../prisma';
+import { platformDB } from '../../prisma';
 
 export async function createOfficialGames(officialGameData: Prisma.OfficialGameCreateManyArgs['data']): Promise<void> {
   await platformDB.officialGame.createMany({
