+++ db/access/player-game.ts
@@ -1,5 +1,5 @@
 import { PlayerGame, Prisma } from '@prisma/client';
-import { platformDB } from '../prisma';
+import { platformDB } from '../../prisma';
 
 export async function createPlayerGame(playerGameData: Prisma.PlayerGameCreateArgs['data']): Promise<PlayerGame> {
   return platformDB.playerGame.create({
