+++ db/access/team.ts
@@ -1,5 +1,5 @@
 import { Team } from '@prisma/client';
-import { platformDB } from '../prisma';
+import { platformDB } from '../../prisma';
 
 export async function getTeamByExternalId(externalId: string): Promise<Team> {
   return platformDB.team.findFirstOrThrow({
