+++ db/access/venue.ts
@@ -1,5 +1,5 @@
 import { Prisma, Venue } from '@prisma/client';
-import { platformDB } from '../prisma';
+import { platformDB } from '../../prisma';
 
 export async function getVenueByExternalId(externalId: string): Promise<Venue | null> {
   return platformDB.venue.findFirst({
